fix: update UserResourceFromEntityAssembler to convert Profile IDs to integers for consistency

+++ services/profiles-service/src/main/java/pe/edu/upc/profiles_service/profiles/interfaces/rest/transform/UserResourceFromEntityAssembler.java
@@ -10,13 +10,13 @@ public class UserResourceFromEntityAssembler {
 
     public static ProfileResource toResourceFromEntity(Profile profile) {
         return new ProfileResource(
-                profile.getId(),
+                profile.getId().intValue(),
                 profile.getName(),
                 profile.getEmail(),
                 profile.getPassword(),
                 profile.getIsActive(),
                 profile.getBirthDate(),
-                profile.getUserProfile().getId()
+                profile.getUserProfile().getId().intValue()
         );
     }
